--- lib/modules/presentation/widgets/custom_exit.dart
@@ -1,43 +1,0 @@
-// ignore_for_file: prefer_const_constructors
-
-import 'package:flutter/material.dart';
-import 'manrope.dart';
-
-class CustomExit extends StatelessWidget {
-  const CustomExit({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return Padding(
-      padding: const EdgeInsets.only(bottom: 60.0),
-      child: Row(
-        children: const [
-          CircleAvatar(
-            radius: 30,
-            // backgroundImage: AssetImage(Helpers.imageJoao),
-          ),
-          Padding(
-            padding: EdgeInsets.only(left: 5.0),
-            child: Column(
-              children: [
-                Manrope(
-                  text: "João",
-                  color: Color.fromARGB(247, 15, 40, 139),
-                  font: FontWeight.w500,
-                  size: 19,
-                ),
-                Manrope(
-                  decoration: TextDecoration.underline,
-                  text: "Sair",
-                  color: Color.fromARGB(247, 15, 40, 139),
-                  font: FontWeight.w500,
-                  size: 15,
-                ),
-              ],
-            ),
-          )
-        ],
-      ),
-    );
-  }
-}
